pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
authors = [
    {name = "Riccardo Paltrinieri",email = "riccardo@paltrinieri.it"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
package-mode = false
dependencies = [
    "google-generativeai",
    "pymongo",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "slack-bolt (>=1.23.0,<2.0.0)",
    "google-cloud-firestore (>=2.20.2,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[tool.black]
line-length = 120
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.optional-dependencies]
dev = [
    "black (>=24.3.0,<25.0.0)",
    "isort (>=5.13.2,<6.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "mypy (>=1.9.0,<2.0.0)",
    "pre-commit (>=3.6.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
